evaluation/requirements.txt
ragas>=0.2.0
arize-phoenix>=5.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
